src/api/requirements.txt
amqp==2.2.1
aniso8601==1.2.1
arrow==0.10.0
beautifulsoup4==4.6.0
billiard==3.5.0.3
cassandra-driver==3.10
celery==4.1.0
click==6.7
django-htmlmin==0.10.0
Flask==0.12.2
Flask-RESTful==0.3.6
futures==3.1.1
html5lib==0.999999999
itsdangerous==0.24
Jinja2==2.9.6
kombu==4.1.0
MarkupSafe==1.0
marshmallow==2.13.5
python-dateutil==2.6.1
pytz==2017.2
scales==1.0.9
six==1.10.0
uWSGI==2.0.15
vine==1.1.4
webencodings==0.5.1
Werkzeug==0.12.2
